+++ tests/testthat/testGeneral.R
@@ -349,6 +349,42 @@ for(i in 1:nrow(scenarios)) {
 }
 
 
+context("ILLEGAL NAME TESTS")
+
+scenarios <- testScenarios("illegal names tests", runAllForReleaseVersion=TRUE)
+illegalNameTestData <- data.frame("Sale ID"=1:5, "Colour"=c("Red", "Red", "Green", "Red", "Green"), "Sale Item"=c("Car", "Lorry", "Car", "Train", "Train"), "Sale Amount"=c(12.1,2.3,5.6,3.7,1.5))
+colnames(illegalNameTestData) <- c("Sale.ID", "Colour Name", "Sale$Item", "Sale+Amount")
+
+for(i in 1:nrow(scenarios)) {
+  evaluationMode <- scenarios$evaluationMode[i]
+  processingLibrary <- scenarios$processingLibrary[i]
+  description <- scenarios$description[i]
+  countFunction <- scenarios$countFunction[i]
+
+  test_that(description, {
+
+    library(pivottabler)
+    pt <- PivotTable$new(processingLibrary=processingLibrary, evaluationMode=evaluationMode)
+    pt$addData(illegalNameTestData)
+    pt$addColumnDataGroups("Sale$Item")
+    pt$addRowDataGroups("Colour Name")
+    pt$defineCalculation(calculationName="Sale Count", summariseExpression=countFunction)
+    pt$defineCalculation(calculationName="$Sale$Count$", summariseExpression=countFunction)
+    pt$defineCalculation(calculationName="Total+Sales", summariseExpression="sum(`Sale+Amount`)")
+    pt$defineCalculation(type="calculation", basedOn=c("$Sale$Count$", "Total+Sales"), format="%.1f",
+                         calculationName="!Avg!Sales!", calculationExpression="values$`Total+Sales`/values$`$Sale$Count$`")
+    pt$evaluatePivot()
+    # pt$renderPivot()
+    # sum(pt$cells$asMatrix(), na.rm=TRUE)
+    # prepStr(as.character(pt$getHtml()))
+    html <- "<table class=\"Table\">\n  <tr>\n    <th class=\"RowHeader\" rowspan=\"2\" colspan=\"1\">&nbsp;</th>\n    <th class=\"ColumnHeader\" colspan=\"4\">Car</th>\n    <th class=\"ColumnHeader\" colspan=\"4\">Lorry</th>\n    <th class=\"ColumnHeader\" colspan=\"4\">Train</th>\n    <th class=\"ColumnHeader\" colspan=\"4\">Total</th>\n  </tr>\n  <tr>\n    <th class=\"ColumnHeader\" colspan=\"1\">Sale Count</th>\n    <th class=\"ColumnHeader\" colspan=\"1\">$Sale$Count$</th>\n    <th class=\"ColumnHeader\" colspan=\"1\">Total+Sales</th>\n    <th class=\"ColumnHeader\" colspan=\"1\">!Avg!Sales!</th>\n    <th class=\"ColumnHeader\" colspan=\"1\">Sale Count</th>\n    <th class=\"ColumnHeader\" colspan=\"1\">$Sale$Count$</th>\n    <th class=\"ColumnHeader\" colspan=\"1\">Total+Sales</th>\n    <th class=\"ColumnHeader\" colspan=\"1\">!Avg!Sales!</th>\n    <th class=\"ColumnHeader\" colspan=\"1\">Sale Count</th>\n    <th class=\"ColumnHeader\" colspan=\"1\">$Sale$Count$</th>\n    <th class=\"ColumnHeader\" colspan=\"1\">Total+Sales</th>\n    <th class=\"ColumnHeader\" colspan=\"1\">!Avg!Sales!</th>\n    <th class=\"ColumnHeader\" colspan=\"1\">Sale Count</th>\n    <th class=\"ColumnHeader\" colspan=\"1\">$Sale$Count$</th>\n    <th class=\"ColumnHeader\" colspan=\"1\">Total+Sales</th>\n    <th class=\"ColumnHeader\" colspan=\"1\">!Avg!Sales!</th>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\" rowspan=\"1\">Green</th>\n    <td class=\"Cell\">1</td>\n    <td class=\"Cell\">1</td>\n    <td class=\"Cell\">5.6</td>\n    <td class=\"Cell\">5.6</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Cell\"></td>\n    <td class=\"Cell\"></td>\n    <td class=\"Cell\"></td>\n    <td class=\"Cell\">1</td>\n    <td class=\"Cell\">1</td>\n    <td class=\"Cell\">1.5</td>\n    <td class=\"Cell\">1.5</td>\n    <td class=\"Cell\">2</td>\n    <td class=\"Cell\">2</td>\n    <td class=\"Cell\">7.1</td>\n    <td class=\"Cell\">3.5</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\" rowspan=\"1\">Red</th>\n    <td class=\"Cell\">1</td>\n    <td class=\"Cell\">1</td>\n    <td class=\"Cell\">12.1</td>\n    <td class=\"Cell\">12.1</td>\n    <td class=\"Cell\">1</td>\n    <td class=\"Cell\">1</td>\n    <td class=\"Cell\">2.3</td>\n    <td class=\"Cell\">2.3</td>\n    <td class=\"Cell\">1</td>\n    <td class=\"Cell\">1</td>\n    <td class=\"Cell\">3.7</td>\n    <td class=\"Cell\">3.7</td>\n    <td class=\"Cell\">3</td>\n    <td class=\"Cell\">3</td>\n    <td class=\"Cell\">18.1</td>\n    <td class=\"Cell\">6.0</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\" rowspan=\"1\">Total</th>\n    <td class=\"Cell\">2</td>\n    <td class=\"Cell\">2</td>\n    <td class=\"Cell\">17.7</td>\n    <td class=\"Cell\">8.8</td>\n    <td class=\"Cell\">1</td>\n    <td class=\"Cell\">1</td>\n    <td class=\"Cell\">2.3</td>\n    <td class=\"Cell\">2.3</td>\n    <td class=\"Cell\">2</td>\n    <td class=\"Cell\">2</td>\n    <td class=\"Cell\">5.2</td>\n    <td class=\"Cell\">2.6</td>\n    <td class=\"Cell\">5</td>\n    <td class=\"Cell\">5</td>\n    <td class=\"Cell\">25.2</td>\n    <td class=\"Cell\">5.0</td>\n  </tr>\n</table>"
+
+    expect_equal(round(sum(pt$cells$asMatrix(), na.rm=TRUE), digits = 3), 194.373)
+    expect_identical(as.character(pt$getHtml()), html)
+  })
+}
+
+
 context("QUICK-PIVOT TESTS")